+++ build_tools/spec/region_spec.rb
@@ -8,7 +8,7 @@ end
 def whitelist
   {
     'core' => {
-      'errors.rb' => 230,
+      'errors.rb' => 240,
       'signature_v4.rb' => 35,
       'stub_responses.rb' => 19
     },