+++ gems/aws-sdk-core/lib/aws-sdk-core/util.rb
@@ -52,6 +52,10 @@ module Aws
         end
       end
 
+      def monotonic_seconds
+        monotonic_milliseconds / 1000.0
+      end
+
       def str_2_bool(str)
         case str.to_s
         when "true" then true